Refactor Socrata classes
