+++ lib/socrata/exporter.rb
@@ -1,16 +1,21 @@
 # frozen_string_literal: true
 
 module Socrata
-  class Exporter
+  class Exporter < OpenData
     class << self
-      # Returns the file name and the data as a String.
+      # Returns both the file name (with timestamp) and the exported data as a String.
       def export(format)
-        file_format = format || default_format
+        log('Exporting dataset...')
 
+        file_format = format || default_format
         export_data = Decidim::Exporters.find_exporter(file_format).new(collection, serializer).export
         file_name = export_data.filename(default_name)
 
+        log("File created: #{file_name}")
+
         [file_name, export_data.read]
+      rescue StandardError => error
+        log_error(error)
       end
 
       private
@@ -21,16 +26,6 @@ module Socrata
 
       def default_name
         'socrata_open_data'
-      end
-
-      def collection
-        Decidim::ParticipatoryProcess.where(
-          organization: Decidim::Organization.first
-        )
-      end
-
-      def serializer
-        Decidim::Process::Extended::ProcessSerializer
       end
     end
   end
