+++ lib/socrata/publisher.rb
@@ -3,22 +3,25 @@
 require 'soda/client'
 
 module Socrata
-  class Publisher
+  class Publisher < OpenData
     class << self
       # Returns a Hashie::Mash that represents the body of the response.
       def publish
-        client.post(identifier, data)
+        log("Pushing data to #{ENV['SODA_DOMAIN']}...")
+
+        response = client.post(dataset_identifier, payload)
+
+        log(response.body)
+      rescue StandardError => error
+        log_error(error)
       end
 
       private
 
-      # Identifier for the dataset we want to access, i.e.
-      # https://soda.demo.socrata.com/dataset/Example-dataset/identifier
-      def identifier
-        'identifier'
-      end
-
+      # Returns a SODA::Client instance.
       # https://www.rubydoc.info/github/socrata/soda-ruby/SODA/Client
+      # Source code:
+      # https://github.com/socrata/soda-ruby/blob/master/lib/soda/client.rb
       def client
         SODA::Client.new(
           domain: ENV['SODA_DOMAIN'],
@@ -28,10 +31,17 @@ module Socrata
         )
       end
 
-      # The soda-ruby post method accepts both JSON and CSV
+      # Identifier for the dataset we want to access, i.e.:
+      # https://soda.demo.socrata.com/dataset/Example-Dataset/identifier
+      def dataset_identifier
+        ENV['SODA_DATASET_IDENTIFIER']
+      end
+
+      # Payload to POST. Will be converted to JSON.
       # http://socrata.github.io/soda-ruby/examples/upsert.html
-      def data
-        Socrata::Exporter.export('JSON').second
+      # Does not accept nil values, so we discard them with Hash.compact().
+      def payload
+        collection.map { |resource| serializer.new(resource).serialize.compact }
       end
     end
   end
