+++ lib/tasks/socrata.rake
@@ -6,20 +6,14 @@ namespace :socrata do
   # - format - The exporter format as a string. Defaults to "CSV"
   # - i.e. rake socrata:export["JSON"]
   task :export, [:format] => :environment do |_task, args|
-    Socrata::Log.log('Exporting dataset...')
-
     file_name, export_data = Socrata::Exporter.export(args.format)
     File.write(file_name, export_data)
-
-    Socrata::Log.log("File created: #{file_name}")
   end
 
-  desc 'Updates Socrata dataset via POST method with soda-ruby'
+  desc 'Updates Socrata dataset via HTTP POST request'
+  # The response object will be a summary of the updates were made
+  # and possibly the errors that were encountered.
   task publish: :environment do
-    Socrata::Log.log('Pushing data to soda.demo.socrata.com...')
-
-    response = Socrata::Publisher.publish
-
-    Socrata::Log.log(response)
+    Socrata::Publisher.publish
   end
 end
